Fix yolo modules
