+++ pyproject.toml
@@ -7,7 +7,8 @@ readme = "README.md"
 packages = [
     { include = "om1_speech", from = "src" },
     { include = "om1_utils", from = "src" },
-    { include = "om1_vlm", from = "src" }
+    { include = "om1_vlm", from = "src" },
+    { include = "om1_ml", from = "src"},
 ]
 
 [tool.poetry.scripts]
@@ -24,6 +25,8 @@ protobuf = "^5.29.3"
 google = "^3.0.0"
 ultralytics = "^8.3.134"
 omegaconf = "^2.3.0"
+flask = "^3.1.1"
+coremltools = "^8.3.0"
 
 [tool.poetry.group.dev.dependencies]
 pytest = "^7.0"
