+++ src/om1_ml/__main__.py
@@ -1,0 +1,4 @@
+from om1_ml.main import main
+
+if __name__ == "__main__":
+    main()
