+++ src/om1_ml/blur/__init__.py
@@ -1,0 +1,3 @@
+from .yolo import YOLOFaceDetection
+
+__all__ = ["YOLOFaceDetection"]
