--- src/om1_vlm/blur/__init__.py
@@ -1,3 +1,0 @@
-from .yolo import YOLOFaceDetection
-
-__all__ = ["YOLOFaceDetection"]
