--- src/om1_vlm/blur/yolov8n-face.mlpackage/Manifest.json
@@ -1,18 +1,0 @@
-{
-    "fileFormatVersion": "1.0.0",
-    "itemInfoEntries": {
-        "13CBBF45-E0F2-43C8-8861-6D258BA8BB9C": {
-            "author": "com.apple.CoreML",
-            "description": "CoreML Model Specification",
-            "name": "model.mlmodel",
-            "path": "com.apple.CoreML/model.mlmodel"
-        },
-        "3A6F91DC-D753-42FC-A64C-75946D78E8A1": {
-            "author": "com.apple.CoreML",
-            "description": "CoreML Model Weights",
-            "name": "weights",
-            "path": "com.apple.CoreML/weights"
-        }
-    },
-    "rootModelIdentifier": "13CBBF45-E0F2-43C8-8861-6D258BA8BB9C"
-}
